+++ src/stuart-ml/util/MLUtils.lua
@@ -27,11 +27,11 @@ M.fastSquaredDistance = function(v1, norm1, v2, norm2, precision)
   local sqDist = 0.0
   local precisionBound1 = 2.0 * M.EPSILON * sumSquaredNorm / (normDiff * normDiff + M.EPSILON)
   local BLAS = require 'stuart-ml.linalg.BLAS'
-  local SparseVector = require 'stuart-ml.linalg.SparseVector'
+  local class = require 'stuart.class'
   local Vectors = require 'stuart-ml.linalg.Vectors'
   if precisionBound1 < precision then
     sqDist = sumSquaredNorm - 2.0 * BLAS.dot(v1, v2)
-  elseif v1:isInstanceOf(SparseVector) or v2:isInstanceOf(SparseVector) then
+  elseif class.istype(v1,'SparseVector') or class.istype(v2,'SparseVector') then
     local dotValue = BLAS.dot(v1, v2)
     sqDist = math.max(sumSquaredNorm - 2.0 * dotValue, 0.0)
     local precisionBound2 = M.EPSILON * (sumSquaredNorm + 2.0 * math.abs(dotValue)) / (sqDist + M.EPSILON)